Solved bug with roulet function. Rounding of double numbers was the issue; double is replaced with decimal.

+++ AlgorithmsCore/Contracts/IAntSystem.cs
@@ -11,7 +11,7 @@
         /// <returns></returns>
         int GetNextColony();
 
-        double[] CalculateProbability(int nextColony);
+        decimal[] CalculateProbability(int nextColony);
         AntSystemFragment UpdatePhermone(double maxAllowedWeight);
     }
 }
